voice_recognition/requirements.txt
pubchempy
requests
numpy 
SpeechRecognition
google-cloud-texttospeech
playsound
pyaudio
PyObjC
git+https://bitbucket.org/mtzcloud/tcc-python.git
